@@ -14,18 +14,21 @@
     "email": "ig.fedorov@corp.vk.com"
   },
   "dependencies": {
+    "@mapbox/hast-util-to-jsx": "^2.1.0",
     "@swc/cli": "^0.6.0",
     "@swc/core": "^1.11.10",
     "glob": "^11.0.1",
-    "svg-baker": "^1.7.0",
+    "hast-util-from-html": "^2.0.3",
     "svgo": "^3.3.2"
   },
   "devDependencies": {
     "typescript": "^5.8.2"
   },
   "engines": {
-    "node": ">12.0.0"
+    "node": ">=20.11.0"
   },
+  "type": "module",
+  "exports": "./index.js",
   "publishConfig": {
     "provenance": true
   }
