@@ -2,16 +2,4 @@
  * @typedef {import('../packages/icons-scripts/scripts/options').DeprecatedIcons} DeprecatedIcons
  * @type DeprecatedIcons
  */
-const DEPRECATED_ICONS = {
-  donate_oultine_16: 'donate_outline_16',
-  info_cirle_16: 'info_circle_16',
-  folder_simple_arrow_right_outlune_20: 'folder_simple_arrow_right_outline_20',
-  apple_watch_outlite_28: 'apple_watch_outline_28',
-  chat_vawe_outline_gray_24: 'chat_wave_outline_gray_24',
-  rectrangle_hand_point_up_24: 'rectangle_hand_point_up_24',
-  phone_vawe_outline_gray_24: 'phone_wave_outline_gray_24',
-};
-
-module.exports = {
-  DEPRECATED_ICONS,
-};
+export const DEPRECATED_ICONS = {};
